refactor: WalletConnect V1
