+++ packages/mobile/package.json
@@ -74,8 +74,12 @@
     "@segment/analytics-react-native-firebase": "^1.4.4",
     "@sentry/react-native": "^2.1.0",
     "@ungap/url-search-params": "^0.1.2",
-    "@walletconnect/client": "2.0.0-beta.17",
-    "@walletconnect/types": "2.0.0-beta.17",
+    "@walletconnect/client-v1": "npm:@walletconnect/client@1.6.5",
+    "@walletconnect/client-v2": "npm:@walletconnect/client@2.0.0-beta.17",
+    "@walletconnect/utils-v1": "npm:@walletconnect/utils@1.6.5",
+    "@walletconnect/utils-v2": "npm:@walletconnect/utils@2.0.0-beta.17",
+    "@walletconnect/types-v1": "npm:@walletconnect/types@1.6.5",
+    "@walletconnect/types-v2": "npm:@walletconnect/types@2.0.0-beta.17",
     "Base64": "^1.0.1",
     "apollo-boost": "^0.4.9",
     "apollo-cache-persist": "^0.1.1",
@@ -230,7 +234,10 @@
     "ts-retry-promise": "^0.6.0"
   },
   "resolutions": {
-    "react-native-fs": "git+https://github.com/celo-org/react-native-fs#aa6db0f"
+    "react-native-fs": "git+https://github.com/celo-org/react-native-fs#aa6db0f", 
+    "@walletconnect/client-v1/@walletconnect-utils": "1.6.5",
+    "@walletconnect/client-v1/@walletconnect-types": "1.6.5",
+    "@walletconnect/utils-v1/@walletconnect-types": "1.6.5"
   },
   "detox": {
     "test-runner": "jest",
